protection with demo

+++ app/Http/Controllers/Customer/ChatBoxController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers\Customer;
 use App\Events\MessageReceived;
 use App\Helpers\Worker;
 use App\Http\Controllers\Controller;
-use App\Models\Blacklists;
 use App\Models\ChatBox;
 use App\Models\ChatBoxMessage;
 use App\Models\Todos;
@@ -18,11 +17,9 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Str;
 use Throwable;
 
-class ChatBoxController extends Controller
-{
+class ChatBoxController extends Controller {
 
     /**
      * get all chat box
@@ -30,28 +27,27 @@ class ChatBoxController extends Controller
      * @return Application|Factory|View
      * @throws AuthorizationException
      */
-    public function open(Todos $todo, ChatBox $box)
-    {
-        $this->authorize('view_chat');
+    public function open( Todos $todo, ChatBox $box ) {
+        $this->authorize( 'view_chat' );
 
-        if (auth()->user()->id != $todo->user_id) {
-            abort(401);
+        if ( auth()->user()->id != $todo->user_id ) {
+            abort( 401 );
         }
 
         $pageConfigs = [
-            'pageHeader'    => false,
+            'pageHeader' => false,
             'contentLayout' => "content-left-sidebar",
-            'bodyClass'     => 'chat-application',
+            'bodyClass' => 'chat-application',
         ];
 
-        $chat_box = ChatBox::where('user_id', Auth::user()->id)->take(1000)->orderBy('updated_at', 'desc')->cursor();
+        $chat_box = ChatBox::where( 'user_id', Auth::user()->id )->take( 1000 )->orderBy( 'updated_at', 'desc' )->cursor();
 
-        return view('customer.chatbox.open', [
+        return view( 'customer.chatbox.open', [
             'pageConfigs' => $pageConfigs,
-            'chat_box'    => $chat_box,
+            'chat_box' => $chat_box,
             'todo' => $todo,
             'box' => $box,
-        ]);
+        ] );
     }
     /**
      * get all chat box
@@ -59,29 +55,27 @@ class ChatBoxController extends Controller
      * @return Application|Factory|View
      * @throws AuthorizationException
      */
-    public function receiver(Todos $todo)
-    {
+    public function receiver( Todos $todo ) {
 
         $pageConfigs = [
-            'pageHeader'    => false,
+            'pageHeader' => false,
             'contentLayout' => "content-left-sidebar",
-            'bodyClass'     => 'chat-application',
+            'bodyClass' => 'chat-application',
         ];
 
-        if (ChatBox::where('to', auth()->user()->id)->count() < 1) {
-            abort(404);
+        if ( ChatBox::where( 'to', auth()->user()->id )->count() < 1 ) {
+            abort( 404 );
         }
 
-        $chat_box = ChatBox::where('to', auth()->user()->id)->take(1000)
-            ->orderBy('updated_at', 'desc')->first();
+        $chat_box = ChatBox::where( 'to', auth()->user()->id )->take( 1000 )
+            ->orderBy( 'updated_at', 'desc' )->first();
 
-        return view('customer.chatbox.receiver.index', [
+        return view( 'customer.chatbox.receiver.index', [
             'pageConfigs' => $pageConfigs,
-            'chat_box'    => $chat_box,
-            'todo' => $todo
-        ]);
+            'chat_box' => $chat_box,
+            'todo' => $todo,
+        ] );
     }
-
 
     /**
      * start new conversation
@@ -91,98 +85,103 @@ class ChatBoxController extends Controller
      * @throws AuthorizationException
      *
      */
-    public function new(Todos $todo)
-    {
-        $this->authorize('create_chat');
+    public function new ( Todos $todo ) {
+        $this->authorize( 'create_chat' );
 
         $breadcrumbs = [
-            ['link' => url('dashboard'), 'name' => __('locale.menu.Dashboard')],
-            ['link' => url('chat-box'), 'name' => __('locale.menu.Chat Box')],
+            ['link' => url( 'dashboard' ), 'name' => __( 'locale.menu.Dashboard' )],
+            ['link' => url( 'chat-box' ), 'name' => __( 'locale.menu.Chat Box' )],
             ['name' => 'add to chat'],
         ];
 
         $users = [];
 
-        $exist_users = ChatBox::where('user_id', Auth::id())
-            ->where('todo_id', $todo->id)->pluck('to');
+        $exist_users = ChatBox::where( 'user_id', Auth::id() )
+            ->where( 'todo_id', $todo->id )->pluck( 'to' );
 
-        if (in_array('all', $todo->assigned())) {
-            $users = User::where('active_portal', 'customer')
-                ->where('id', '!=', auth()->user()->id)
-                ->whereNotIn('id', $exist_users)
+        if ( in_array( 'all', $todo->assigned() ) ) {
+            $users = User::where( 'active_portal', 'customer' )
+                ->where( 'id', '!=', auth()->user()->id )
+                ->whereNotIn( 'id', $exist_users )
                 ->get();
         } else {
-            foreach ($todo->assigned() as $user_id) {
-                User::where('id', $user_id)->whereNotIn('id', $exist_users);
-                if (User::where('id', $user_id)->whereNotIn('id', $exist_users)->count() > 0) {
-                    $users[] = User::where('id', $user_id)->whereNotIn('id', $exist_users)->first();
+            foreach ( $todo->assigned() as $user_id ) {
+                User::where( 'id', $user_id )->whereNotIn( 'id', $exist_users );
+                if ( User::where( 'id', $user_id )->whereNotIn( 'id', $exist_users )->count() > 0 ) {
+                    $users[] = User::where( 'id', $user_id )->whereNotIn( 'id', $exist_users )->first();
                 }
             }
         }
 
-        $chatbox = ChatBox::where('user_id', auth()->user()->id)->where('todo_id', $todo->id)->get();
+        $chatbox = ChatBox::where( 'user_id', auth()->user()->id )->where( 'todo_id', $todo->id )->get();
 
-        return view('customer.chatbox.new', compact('breadcrumbs', 'users', 'todo'));
+        return view( 'customer.chatbox.new', compact( 'breadcrumbs', 'users', 'todo' ) );
     }
-
 
     /**
      * start new conversion
      * @param  Request  $request
      */
-    public function store(Request $request)
-    {
-        $this->authorize('create_chat');
+    public function store( Request $request ) {
 
-        $request->validate([
+        if ( $this->checks() ) {
+            return response()->json( [
+                'status' => 'error',
+                'message' => 'Sorry! This option is not available in demo mode',
+            ] );
+        }
+
+        $this->authorize( 'create_chat' );
+
+        $request->validate( [
             'todo_id' => 'required|exists:todos,id',
             'user_id' => 'required|exists:users,id',
-        ]);
+        ] );
 
-        $todo = Todos::find($request->todo_id);
-        $user = User::find($request->user_id);
+        $todo = Todos::find( $request->todo_id );
+        $user = User::find( $request->user_id );
 
-        $chatbox = ChatBox::where('user_id', Auth::user()->id)
-            ->where('todo_id', $request->todo_id)
-            ->where('to', $request->user_id)->count();
+        $chatbox = ChatBox::where( 'user_id', Auth::user()->id )
+            ->where( 'todo_id', $request->todo_id )
+            ->where( 'to', $request->user_id )->count();
 
-        if ($chatbox > 0) {
-            return redirect()->route('customer.chat.new', $todo->uid)->with([
-                'status'  => 'error',
-                'message' => __('Already exists chat with this user!'),
-            ]);
+        if ( $chatbox > 0 ) {
+            return redirect()->route( 'customer.chat.new', $todo->uid )->with( [
+                'status' => 'error',
+                'message' => __( 'Already exists chat with this user!' ),
+            ] );
         }
 
         $chatbox = new ChatBox();
 
-        $createBox = $chatbox->create([
-            'user_id'           => Auth::user()->id,
-            'from'              => auth()->user()->id,
-            'to'                => $request->user_id,
-            'todo_id'           => $todo->id,
-            'notification'      => 1,
-        ]);
+        $createBox = $chatbox->create( [
+            'user_id' => Auth::user()->id,
+            'from' => auth()->user()->id,
+            'to' => $request->user_id,
+            'todo_id' => $todo->id,
+            'notification' => 1,
+        ] );
 
-        if ($createBox->save()) {
+        if ( $createBox->save() ) {
 
-            ChatBoxMessage::create([
-                'box_id'  => $createBox->id,
+            ChatBoxMessage::create( [
+                'box_id' => $createBox->id,
                 'message' => 'Task: ' . $todo->name,
                 'send_by' => 'from',
-            ]);
+            ] );
 
             $createBox->touch();
 
-            return redirect()->route('customer.chat.opne_with_user', [$todo->uid, $createBox->uid])->with([
-                'status'  => 'success',
-                'message' => __('chat box has been created successfully'),
-            ]);
+            return redirect()->route( 'customer.chat.opne_with_user', [$todo->uid, $createBox->uid] )->with( [
+                'status' => 'success',
+                'message' => __( 'chat box has been created successfully' ),
+            ] );
         }
 
-        return redirect()->route('customer.chat.new', $todo->uid)->with([
-            'status'  => 'error',
-            'message' => __('unale to create chat box'),
-        ]);
+        return redirect()->route( 'customer.chat.new', $todo->uid )->with( [
+            'status' => 'error',
+            'message' => __( 'unale to create chat box' ),
+        ] );
     }
 
     /**
@@ -192,18 +191,17 @@ class ChatBoxController extends Controller
      *
      * @return JsonResponse
      */
-    public function messages(ChatBox $box): JsonResponse
-    {
-        $box->update([
+    public function messages( ChatBox $box ): JsonResponse {
+        $box->update( [
             'notification' => 0,
-        ]);
+        ] );
 
-        $data = ChatBoxMessage::where('box_id', $box->id)->select('message', 'send_by', 'media_url')->cursor()->toJson();
+        $data = ChatBoxMessage::where( 'box_id', $box->id )->select( 'message', 'send_by', 'media_url' )->cursor()->toJson();
 
-        return response()->json([
+        return response()->json( [
             'status' => 'success',
-            'data'   => $data,
-        ]);
+            'data' => $data,
+        ] );
     }
 
     /**
@@ -213,16 +211,14 @@ class ChatBoxController extends Controller
      *
      * @return JsonResponse
      */
-    public function messagesWithNotification(ChatBox $box): JsonResponse
-    {
-        $data = ChatBoxMessage::where('box_id', $box->id)->select('message', 'send_by', 'media_url')->cursor()->toJson();
+    public function messagesWithNotification( ChatBox $box ): JsonResponse {
+        $data = ChatBoxMessage::where( 'box_id', $box->id )->select( 'message', 'send_by', 'media_url' )->cursor()->toJson();
 
-        return response()->json([
+        return response()->json( [
             'status' => 'success',
-            'data'   => $data,
-        ]);
+            'data' => $data,
+        ] );
     }
-
 
     /**
      * reply message
@@ -230,43 +226,47 @@ class ChatBoxController extends Controller
      * @param  ChatBox  $box
      * @param  Request  $request
      */
-    public function reply(ChatBox $box, Request $request)
-    {
+    public function reply( ChatBox $box, Request $request ) {
 
-        $this->authorize('create_chat');
+        if ( $this->checks() ) {
+            return response()->json( [
+                'status' => 'error',
+                'message' => 'Sorry! This option is not available in demo mode',
+            ] );
+        }
 
-        if (empty($request->message)) {
-            return response()->json([
-                'status'  => 'error',
-                'message' => __('locale.campaigns.insert_your_message'),
-            ]);
+        $this->authorize( 'create_chat' );
+
+        if ( empty( $request->message ) ) {
+            return response()->json( [
+                'status' => 'error',
+                'message' => __( 'locale.campaigns.insert_your_message' ),
+            ] );
         }
 
         try {
 
-            $message = Worker::linkify($request->message);
+            $message = Worker::linkify( $request->message );
 
-            ChatBoxMessage::create([
-                'box_id'  => $box->id,
-                'message' =>  $message,
+            ChatBoxMessage::create( [
+                'box_id' => $box->id,
+                'message' => $message,
                 'send_by' => 'from',
-            ]);
+            ] );
 
             $box->touch();
 
-
-            event(new MessageReceived(auth()->user(), $message, [
-                'recipient_id' => $box->to
-            ]));
-
+            event( new MessageReceived( auth()->user(), $message, [
+                'recipient_id' => $box->to,
+            ] ) );
 
             return 'sent';
             ///s
-        } catch (Throwable $th) {
-            return response()->json([
-                'status'  => 'error',
+        } catch ( Throwable $th ) {
+            return response()->json( [
+                'status' => 'error',
                 'message' => $th->getMessage(),
-            ]);
+            ] );
         }
     }
     /**
@@ -275,43 +275,48 @@ class ChatBoxController extends Controller
      * @param  ChatBox  $box
      * @param  Request  $request
      */
-    public function replyTo(ChatBox $box, Request $request)
-    {
+    public function replyTo( ChatBox $box, Request $request ) {
 
-        $this->authorize('create_chat');
+        if ( $this->checks() ) {
+            return response()->json( [
+                'status' => 'error',
+                'message' => 'Sorry! This option is not available in demo mode',
+            ] );
+        }
 
-        if (empty($request->message)) {
-            return response()->json([
-                'status'  => 'error',
-                'message' => __('locale.campaigns.insert_your_message'),
-            ]);
+        $this->authorize( 'create_chat' );
+
+        if ( empty( $request->message ) ) {
+            return response()->json( [
+                'status' => 'error',
+                'message' => __( 'locale.campaigns.insert_your_message' ),
+            ] );
         }
 
         try {
 
-            $message = Worker::linkify($request->message);
+            $message = Worker::linkify( $request->message );
 
-            ChatBoxMessage::create([
-                'box_id'  => $box->id,
+            ChatBoxMessage::create( [
+                'box_id' => $box->id,
                 'message' => $message,
                 'send_by' => 'to',
-            ]);
+            ] );
 
-            event(new MessageReceived(auth()->user(), $message, [
-                'recipient_id' => $box->user_id
-            ]));
+            event( new MessageReceived( auth()->user(), $message, [
+                'recipient_id' => $box->user_id,
+            ] ) );
 
             $box->touch();
 
             return 'sent';
-        } catch (Throwable $th) {
-            return response()->json([
-                'status'  => 'error',
+        } catch ( Throwable $th ) {
+            return response()->json( [
+                'status' => 'error',
                 'message' => $th->getMessage(),
-            ]);
+            ] );
         }
     }
-
 
     /**
      * delete chatbox messages
@@ -320,22 +325,28 @@ class ChatBoxController extends Controller
      *
      * @return JsonResponse
      */
-    public function delete(ChatBox $box): JsonResponse
-    {
-        $messages = ChatBoxMessage::where('box_id', $box->id)->delete();
-        if ($messages) {
+    public function delete( ChatBox $box ): JsonResponse {
+
+        if ( $this->checks() ) {
+            return response()->json( [
+                'status' => 'error',
+                'message' => 'Sorry! This option is not available in demo mode',
+            ] );
+        }
+        $messages = ChatBoxMessage::where( 'box_id', $box->id )->delete();
+        if ( $messages ) {
             $box->delete();
 
-            return response()->json([
-                'status'  => 'success',
-                'message' => __('message deleted'),
-            ]);
+            return response()->json( [
+                'status' => 'success',
+                'message' => __( 'message deleted' ),
+            ] );
         }
 
-        return response()->json([
-            'status'  => 'error',
-            'message' => __('locale.exceptions.something_went_wrong'),
-        ]);
+        return response()->json( [
+            'status' => 'error',
+            'message' => __( 'locale.exceptions.something_went_wrong' ),
+        ] );
     }
 
     /**
@@ -345,11 +356,10 @@ class ChatBoxController extends Controller
      *
      * @return JsonResponse
      */
-    public function block(): JsonResponse
-    {
-        return response()->json([
-            'status'  => 'error',
-            'message' => __('under development'),
-        ]);
+    public function block(): JsonResponse {
+        return response()->json( [
+            'status' => 'error',
+            'message' => __( 'under development' ),
+        ] );
     }
 }
